Release Selene 6.2.0 with native Moonlight OS integration

+++ scripts/build-appimage.sh
@@ -22,6 +22,12 @@ fi
 
 command -v qmake6 >/dev/null 2>&1 || fail "Unable to find 'qmake6' in your PATH!"
 command -v $LINUXDEPLOY >/dev/null 2>&1 || fail "Unable to find '$LINUXDEPLOY' in your PATH!"
+[ -f "$MSQUIC_LIBRARY" ] || fail "MSQUIC_LIBRARY must point to the pinned MsQuic runtime!"
+# linuxdeploy resolves the executable's DT_NEEDED entries before it processes
+# --library arguments. Make the extracted runtime's SONAME visible during
+# that first dependency scan; --library below still copies it into the image.
+MSQUIC_LIBRARY_DIR=$(dirname "$MSQUIC_LIBRARY")
+export LD_LIBRARY_PATH="$MSQUIC_LIBRARY_DIR${LD_LIBRARY_PATH:+:$LD_LIBRARY_PATH}"
 
 echo Cleaning output directories
 rm -rf $BUILD_FOLDER
@@ -46,7 +52,7 @@ pushd $BUILD_FOLDER
 # work even in X11. To avoid this, we will disable Wayland support for the AppImage.
 #
 # We disable DRM support because linuxdeploy doesn't bundle the appropriate libraries for Qt EGLFS.
-qmake6 $SOURCE_ROOT/selene.pro CONFIG+=disable-wayland CONFIG+=disable-libdrm PREFIX=$DEPLOY_FOLDER/usr DEFINES+=APP_IMAGE || fail "Qmake failed!"
+qmake6 $SOURCE_ROOT/selene.pro CONFIG+=disable-wayland CONFIG+=disable-libdrm PREFIX=$DEPLOY_FOLDER/usr DEFINES+=APP_IMAGE MSQUIC_LIBRARY="$MSQUIC_LIBRARY" || fail "Qmake failed!"
 popd
 
 echo Compiling Selene in $BUILD_CONFIG configuration
@@ -66,6 +72,7 @@ echo Creating AppImage
 pushd $INSTALLER_FOLDER
 VERSION=$VERSION $LINUXDEPLOY --appdir $DEPLOY_FOLDER \
   --library=/usr/local/lib/libSDL3.so.0 \
+  --library="$MSQUIC_LIBRARY" \
   --plugin qt --output appimage || fail "linuxdeploy failed!"
 popd
